backend/requirements.txt
# FastAPI e dependências principais
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Banco de dados
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Redis
redis[hiredis]==5.0.1

# Autenticação e segurança
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8

# Validação e serialização
pydantic[email]==2.5.0
email-validator==2.1.0

# HTTP requests
httpx==0.25.2
requests==2.31.0

# WhatsApp e comunicação
twilio==8.10.3
sendgrid==6.10.0

# Utilitários
python-dateutil==2.8.2
pytz==2023.3
uuid==1.30

# Desenvolvimento e testes
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
flake8==6.1.0

# Monitoramento e logging
structlog==23.2.0
sentry-sdk[fastapi]==1.38.0

# Upload de arquivos
python-magic==0.4.27
pillow==10.1.0

# Rate limiting
slowapi==0.1.9

# Configuração
pydantic-settings==2.1.0
